Set parameters of Animal, Create class Move

+++ src/ru/javarush/maxzaharov/animal_island/plants/Plant.java
@@ -7,6 +7,6 @@ import java.util.ArrayList;
 
 
 public class Plant extends BasicUnit {
-    public int weight = 1;
+    public double weight = 1;
 
    }
